requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
PyJWT==2.8.0
cryptography==42.0.8
google-generativeai==0.7.1
webbrowser==1.0
requests==2.32.3
google-auth==2.30.0